Update + Debug Issue
So I tried to fix the KMP function by removing text_index++ whenever a mismatch occured, but there is an issue.
(For inputs: "AABAACAADAABAABA" "AABA")
By the time text_index got to 5 (that is the value C), and pattern_index was 2, the mismatch occured and there was a loop where each previous index of pattern is called without text increasing, which got to the point LPS[0 - 1], which is an impossible number. What should be done in this situation? I am unsure.